moved stuff around and removed go.mod from swag_gen
