+++ backend/internal/bumbal/bumbal_swag_models.go
@@ -1,0 +1,34 @@
+package bumbal
+
+// this file contains models generated by swagger from bumbal's swagger.json
+// but they are altered to only include what we need/ fix some missmatching types
+
+// ZoneModel struct for ZoneModel
+// type BumbalZoneModel struct {
+// 	// Unique ID
+// 	Id string `json:"id,omitempty"`
+// 	// Zone Name
+// 	Name *string `json:"name,omitempty"`
+// 	// Tag names
+// 	FilterTagNames *[]string `json:"filter_tag_names,omitempty"`
+// 	// 
+// 	FilterTags *map[string]interface{} `json:"filter_tags,omitempty"`
+// 	// 
+// 	Links *[]LinkModel `json:"links,omitempty"`
+// 	// 
+// 	MetaData *[]MetaDataModel `json:"meta_data,omitempty"`
+// 	// 
+// 	ZoneRanges *[]ZoneRangeModel `json:"zone_ranges,omitempty"`
+// 	// 
+// 	Brands *[]BrandModel `json:"brands,omitempty"`
+// 	// Brand ID's
+// 	BrandIds *[]int32 `json:"brand_ids,omitempty"`
+// 	// created_at date time
+// 	ZoneCreatedAt *time.Time `json:"zone_created_at,omitempty"`
+// 	// updated_at date time
+// 	ZoneUpdatedAt *time.Time `json:"zone_updated_at,omitempty"`
+// 	// created_by user id
+// 	ZoneCreatedBy *int32 `json:"zone_created_by,omitempty"`
+// 	// created_by user id
+// 	ZoneUpdatedBy *int32 `json:"zone_updated_by,omitempty"`
+// }
