--- backend/internal/swag_gen/go.mod
@@ -1,7 +1,0 @@
-module github.com/GIT_USER_ID/GIT_REPO_ID
-
-go 1.13
-
-require (
-	golang.org/x/oauth2 v0.0.0-20190604053449-0f29369cfe45
-)
